moved filtering code to own module
